back-end/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.2.0
h11==0.11.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kaitaistruct==0.9
ldap3==2.8.1
lxml==4.6.2
MarkupSafe==1.1.1
mitmproxy==5.3.0
msgpack==1.0.0
passlib==1.7.4
protobuf==3.13.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
pyee==7.0.4
PyJWT==1.7.1
pymongo==3.11.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
pyppeteer==0.2.2
python-dotenv==0.15.0
retrying==1.3.3
ruamel.yaml==0.16.12
selenium==3.141.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SQLAlchemy==1.3.20
SQLAlchemy-serializer==1.3.4.4
tornado==6.1
tqdm==4.54.1
urllib3==1.26.2
urwid==2.1.2
websockets==8.1
Werkzeug==1.0.1
wsproto==0.15.0
zstandard==0.14.1
